+++ packages/epetra/test/Map_LL/checkmap.h
@@ -45,6 +45,6 @@
 #include "Epetra_Map.h"
 #include "../epetra_test_err.h"
 int checkmap(Epetra_Map & Map, long long NumGlobalElements, int NumMyElements,
-	     long long * MyGlobalElements, int IndexBase, Epetra_Comm & Comm,
+	     long long * MyGlobalElements, long long IndexBase, Epetra_Comm & Comm,
           bool DistributedGlobal);
